requirements.txt
# Core Django (compatible with Python 3.8)
Django==4.2.16
asgiref==3.6.0
sqlparse==0.4.4
backports.zoneinfo;python_version<"3.9"

# Django REST Framework
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2

# CORS handling
django-cors-headers==4.0.0

# Database
psycopg2-binary==2.9.5
dj-database-url==2.0.0

# Environment management
django-environ==0.10.0
python-decouple==3.7

# Security and Cryptography
cryptography==38.0.4
PyJWT==2.6.0

# Development tools
django-extensions==3.2.1
django-widget-tweaks==1.4.12

# Static files and media
whitenoise==6.4.0
Pillow==9.5.0

# Server
gunicorn==20.1.0
uvicorn==0.20.0

# Utilities
python-dateutil==2.8.2
pytz==2022.7.1

# Testing
pytest==7.2.2
pytest-django==4.5.2

# Optional: For better development experience
ipython==8.12.3
django-debug-toolbar==4.0.0
